requirements.txt
google-cloud-aiplatform
fastapi
uvicorn
google-generativeai